Favorites Assignment

+++ app/controllers/favorites_controller.rb
@@ -3,6 +3,7 @@ class FavoritesController < ApplicationController
 	def create
 		@post = Post.find(params[:post_id])
 		favorite = current_user.favorites.build(post: @post)
+		authorize favorite
 
 		if favorite.save
 			flash[:notice] = "Favorited!"
@@ -14,8 +15,9 @@ class FavoritesController < ApplicationController
   end
 
   def destroy
-  	@post = Post.find(params[:id])
+  	@post = Post.find(params[:post_id])
   	favorite = @post.favorites.find(params[:id])
+  	authorize favorite
 
   	if favorite.destroy
   		flash[:notice] = "Favorite was removed."
